Add PicklerOperations for mscorlib and ManifestModuleRef (#10)
This results in way less bytes for two common cases. Firstly references to mscorlib are frequent and now that's just a single byte token. Secondly most modules are the manifest module of the assembly so we don't need their name, just write out a single byte token saying its the manifest module.