Don't use field type as static types (#14)
A type's fields may change between serialisation time and deserialisation time. We can pretty
easily catch errors where fields have been removed or added or changed name. Catching a change of type is harder.
The way we're doing it now is we've split off some memoisation logic from the current static type, instead passing around a tuple of (Static type * should memoise). For fields of types this lets us check the field type to see if it would be a memosied field, but then set the static type to `object` so that no assumptions are made about the type when reading.

e.g. if the field was an Int32 and we pass that as a static type we would just write the 4 bytes of the Int32, if the field was then changed to an Int64 and we read it with that static type we'd read the 4 bytes written by the Int32 and then the next random 4 bytes of the pickle stream!