Fix race condition with multiple seperate deserialisations
We might have multiple deserialisations happenning at the same time (for example during our tests). Because we throw in the handler if we can't find the type this could result in one deserialisation throwing for a lookup that another deserialisation was trying to do.

This PR changes it so we have one type lookup handler which iterates through all active deserialisation states to do the type lookup. Only throwing if we can't find the assembly that was being looked up (N.B. It might be ok if we just do nothing and let later handlers try to deal with this as well).