Disable changes from reference types to value types (and vice versa)
We'll want to see about trying to support this as best we can at some point in the future, but
it's more tricky than we origionally thought.

The big problem is due to cycles in object graphs when those objects are then changed to value types.
e.g. given a ref type X with a field "Foo : X", you could make a cyclic graph like:
```
var x = new X();
x.Foo = x;
```

This is fine if X is a reference type, but this is impossible to re-construct if X is a value type.
However there are other cycles where it would be possible, for example if Foo was an object not an X,
then it would be a boxed reference to itself which I think we could resolve.

But this is all pretty edge case and its much simpler to just error out for now.